Check if password fields are empty
Before, the <p> element with the requirement class would display 'Passwords
match' if both input fields were empty.

Added a condition inside of the password_validation method that checks if
the values of the inputs are equal to empty strings
If they are, then the input fields and requirement text turn red and a message
stating that 'Password fields are empty' is shown on the webpage.